backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
neo4j==5.14.1
nltk==3.8.1